requirements.txt
streamlit
pdfplumber
pandas
openpyxl
msal
python-dotenv